+++ docs/CHANGELOG.rst
@@ -5,6 +5,16 @@ Change Log
 All notable changes to this project are documented in this file.
 
 
+==========
+Unreleased
+==========
+
+Changed
+-------
+- **BACKWARD INCOMPATIBLE:** Remove ``inherit_collection`` parameter in
+  ``Data.duplicate()`` and ``Sample.duplicate()``
+
+
 ===================
 20.0.0 - 2023-10-27
 ===================