+++ docs/CHANGELOG.rst
@@ -31,6 +31,9 @@ Fixed
 - Fix download stopped with incomplete data received: urllib3 version 2.0 has 
   ``enforce_content_length`` set to ``True`` by default which raises error
   if not enough data was received
+- Require ``urllib`` version ``2.0`` or higher only for ``Python 3.10+`` due to
+  compatibility issues with ``Python 3.9`` or lower
+
 
 ===================
 20.0.0 - 2023-10-27