+++ docs/CHANGELOG.rst
@@ -5,9 +5,9 @@ Change Log
 All notable changes to this project are documented in this file.
 
 
-==========
-Unreleased
-==========
+===================
+21.0.0 - 2023-01-15
+===================
 
 Changed
 -------